+++ package.json
@@ -9,13 +9,16 @@
     "prettier": "prettier --write ."
   },
   "dependencies": {
+    "@moefe/vue-aplayer": "^2.0.0-beta.5",
     "@types/node": "^17.0.22",
-    "@vue/composition-api": "^1.4.9",
+    "aplayer": "^1.10.1",
     "axios": "^0.26.1",
     "element-plus": "^2.0.5",
     "prettier": "^2.6.0",
     "vue": "^3.2.25",
-    "vue-router": "^4.0.13"
+    "vue-aplayer": "^1.6.1",
+    "vue-router": "^4.0.13",
+    "vue3-video-play": "^1.3.1-beta.6"
   },
   "devDependencies": {
     "@vitejs/plugin-vue": "^2.0.0",
