+++ src/api/skins.ts
@@ -79,7 +79,6 @@ export const getAllShare = (params: {
   // 不加params：get的参数传不进去，即不会/share?num=X&size=X&这样请求
   pageNum: number;
   pageSize: number;
-  type: string;
   search: string;
 }) => {
   return http({
