+++ src/env.d.ts
@@ -1,8 +1,13 @@
 /// <reference types="vite/client" />
 
-declare module '*.vue' {
-  import { DefineComponent } from 'vue'
+declare module "*.vue" {
+  import { DefineComponent } from "vue";
   // eslint-disable-next-line @typescript-eslint/no-explicit-any, @typescript-eslint/ban-types
-  const component: DefineComponent<{}, {}, any>
-  export default component
+  const component: DefineComponent<{}, {}, any>;
+  export default component;
 }
+
+declare module "vue3-video-play";
+declare module "vue-aplayer";
+declare module "APlayer";
+declare module "@moefe/vue-aplayer";
